will ADD: add group description on individual group page, it can also be edited. this feature is refactored into its own component
